python/pyproject.toml
[project]
name = "diary-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-google-genai>=4.2.0",
    "openpyxl>=3.1.5",
    "psycopg[binary]>=3.3.4",
    "python-dotenv>=1.2.1",
]
